requirements.txt
streamlit==1.43.2
python-dotenv==1.0.1
google-api-python-client==2.169.0
yt-dlp==2025.2.19
